+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+# 1.0.4 (2018-08-10)
+
+## Bug Fixes
+* Dependencies updated
+
+## Features
+  * Ngx-bootstrap replaced by Angular Material
+
+# 1.0.3 (2018-08-06)
+
+## Bug Fixes
+* Dependencies updated
+
+## Features
+  * Login bug fixed
+  * Emails being displayed for the admin
+
 # 1.0.2 (2018-07-30)
 
 ## Bug Fixes